Добавил data.js, импортировал создание комментариев. Добавил модуль
